refactor: remove most schema refs
