+++ src/lib/openapi/spec/create-feature-schema.ts
@@ -1,4 +1,4 @@
-import { createSchemaObject, CreateSchemaType } from '../types';
+import { createSchemaObject, CreateSchemaType } from '../schema';
 
 const schema = {
     type: 'object',
@@ -17,7 +17,6 @@ const schema = {
             type: 'boolean',
         },
     },
-    'components/schemas': {},
 } as const;
 
 export type CreateFeatureSchema = CreateSchemaType<typeof schema>;
