+++ src/lib/openapi/spec/feature-strategy-schema.ts
@@ -1,4 +1,4 @@
-import { createSchemaObject, CreateSchemaType } from '../types';
+import { createSchemaObject, CreateSchemaType } from '../schema';
 import { constraintSchema } from './constraint-schema';
 import { parametersSchema } from './parameters-schema';
 
@@ -42,17 +42,9 @@ export const schema = {
         },
         constraints: {
             type: 'array',
-            items: {
-                $ref: '#/components/schemas/constraintSchema',
-            },
+            items: constraintSchema,
         },
-        parameters: {
-            $ref: '#/components/schemas/parametersSchema',
-        },
-    },
-    'components/schemas': {
-        constraintSchema,
-        parametersSchema,
+        parameters: parametersSchema,
     },
 } as const;
