+++ src/lib/openapi/spec/features-schema.ts
@@ -1,12 +1,5 @@
-import { createSchemaObject, CreateSchemaType } from '../types';
+import { createSchemaObject, CreateSchemaType } from '../schema';
 import { featureSchema } from './feature-schema';
-import { parametersSchema } from './parameters-schema';
-import { variantSchema } from './variant-schema';
-import { overrideSchema } from './override-schema';
-import { featureEnvironmentSchema } from './feature-environment-schema';
-import { featureStrategySchema } from './feature-strategy-schema';
-import { constraintSchema } from './constraint-schema';
-import { strategySchema } from './strategy-schema';
 
 const schema = {
     type: 'object',
@@ -18,20 +11,8 @@ const schema = {
         },
         features: {
             type: 'array',
-            items: {
-                $ref: '#/components/schemas/featureSchema',
-            },
+            items: featureSchema,
         },
-    },
-    'components/schemas': {
-        featureSchema,
-        constraintSchema,
-        featureEnvironmentSchema,
-        featureStrategySchema,
-        overrideSchema,
-        parametersSchema,
-        strategySchema,
-        variantSchema,
     },
 } as const;
